Switching termdash test to common spinlock utility.
And delaying redraws to allow fast widgets to process the input event.
